Moves search components to Search
